listed changes in desc.
mod.periodic() is called in SwerveSubsystem, move it from the isChassisMoving() to the top of the SwerveSubsystem's periodic (because it should only be called once per robot cycle so if someone else calls your isChassisMoving that it would be called an extra time)

 The drive function in SwerveSubsystem should use getHeading() instead of getYaw() (all the other places that getYaw is used are good)

Add resetToAbsolute() right before burnFlash() in SwerveModule. We do it automatically if the robot is not moving but would be good to do it once at boot up as a backup.